+++ Reach/test/utopia/reach/test/ReachComponentTest.scala
@@ -92,7 +92,7 @@ object ReachComponentTest extends App
 	// Adds user interaction
 	val label = window.result
 	KeyboardEvents += KeyTypedListener.unconditional { event => label.text += event.typedChar.toString }
-	window.focusKeyStateHandler += KeyStateListener(BackSpace) { _ =>
+	window.focusKeyStateHandler += KeyStateListener.unconditional(BackSpace) { _ =>
 		label.text = label.text.string.drop(1).noLanguageLocalizationSkipped
 	}
 }