+++ Paradigm/Changes.md
@@ -30,6 +30,7 @@
   **LinearTransformable** and **AffineTransformable**.
 - Slightly refactored **RelativePoint** mapping logic
 - **Bounds**`.relativize(...)` now accepts any kind of **DoubleVector** instead of just a **Point**
+- Added `.toString` to **RelativePoint**
 
 ## v1.5 - 22.01.2024
 This update contains some major refactoring in the following classes / concepts: